requirements.txt
ragas>=0.0.22
boto3>=1.34.0
azure-openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
langchain>=0.1.0
ragchecker>=0.1.0
pytest>=7.0.0
pydantic>=2.0.0 